% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binaryCodingCpp}
\alias{binaryCodingCpp}
\title{binary (XOR) distance coding for char and bool factors}
\usage{
binaryCodingCpp(data, w)
}
\arguments{
\item{data}{data frame}
}
\description{
binary (XOR) distance coding for char and bool factors
}
