% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{Cookie-class}
\alias{Cookie-class}
\alias{Cookie}
\title{The Cookie Data object Class}
\description{
The Cookie Data object is used to store the all the data, process and results.
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{Raw data frame}

\item{\code{factor.type}}{Types of factor}

\item{\code{normalize.data}}{Normalized data frame}

\item{\code{dist.matrix}}{Distance matrix}

\item{\code{reduction}}{A list of dimmensional reduction objects for this object}

\item{\code{sample.size.test}}{A list of sample size initial test for this object}

\item{\code{samplings}}{A list of sampling results}

\item{\code{version}}{Version of Cookie this object was built under}
}}

