% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{createCookieObject}
\alias{createCookieObject}
\title{Create a Cookie object}
\usage{
createCookieObject(data = NULL, factor.type = NULL)
}
\arguments{
\item{data}{Data frame of raw data.}

\item{factor.type}{Types of each factor. Can be "bool","char","num"}
}
\description{
Create a Cookie object from a data frame
}
