% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{createDimReductionObject}
\alias{createDimReductionObject}
\title{Create a DimReduction object}
\usage{
createDimReductionObject(data = NULL, method = NULL)
}
\arguments{
\item{method}{Method generate this embedding, could be t-SNE, UMAP or MDS}

\item{embedding}{Data frame of raw data.}
}
\description{
Create a DimReduction object from a data frame
}
