% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{createSampleSizeTestObject}
\alias{createSampleSizeTestObject}
\title{Create a SampleSizeTest object}
\usage{
createSampleSizeTestObject(prime.factor = NULL, coverage = NULL,
  selection = NULL, coveragecc = NULL)
}
\arguments{
\item{prime.factor}{Prime factor}

\item{coverage}{Coverage}

\item{selection}{Method generate this embedding, could be t-SNE, UMAP or MDS}
}
\description{
Create a SampleSizeTest object from a data frame
}
