% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{createSamplingObject}
\alias{createSamplingObject}
\title{Create a Sampling object}
\usage{
createSamplingObject(prime.factor = NULL, important.factor = NULL,
  size = NULL, sampling = NULL, coverage = NULL, coveragecc = NULL)
}
\arguments{
\item{prime.factor}{Prime factor}

\item{important.factor}{Important factors}

\item{size}{Sample size}

\item{sampling}{Sampling results}

\item{coverage}{Coverage}
}
\description{
Create a Sampling object from a data frame
}
