% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{DimReduction-class}
\alias{DimReduction-class}
\alias{DimReduction}
\title{The Dimmension Reduction Class}
\description{
The DimReduction object stores a dimension reduction result generated by t-SNE, UMAP or MDS
}
\section{Slots}{

\describe{
\item{\code{embedding}}{Embeddings data frame (required)}

\item{\code{method}}{Method this embedding was built under}
}}

