% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{distCalculation}
\alias{distCalculation}
\title{distCalculation}
\usage{
distCalculation(object = NULL, weight = NULL)
}
\arguments{
\item{object}{Cookie object}

\item{weight}{weight for each factor}
}
\description{
calculate distance from normalized data
}
