% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getSampling}
\alias{getSampling}
\title{getSampling}
\usage{
getSampling(object = NULL, name = NULL)
}
\arguments{
\item{object}{Cookie object}

\item{name}{Name for a specific sampling run. e.g. run1}
}
\description{
Get selected samples from a sampling
}
