% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_cookie.R
\name{give_cookie}
\alias{give_cookie}
\title{Give a cookie to someone}
\usage{
give_cookie(name)
}
\arguments{
\item{name}{a character defining the name of someon you want to give a cookie to}
}
\value{
what kind of cookie they got
}
\description{
Give a cookie to someone
}
\examples{
give_cookie("Zhian")
}
