% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hammingCodingCpp}
\alias{hammingCodingCpp}
\title{hamming distance coding for numerical factors}
\usage{
hammingCodingCpp(data)
}
\arguments{
\item{data}{data frame}
}
\description{
hamming distance coding for numerical factors
}
