% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotDistribution}
\alias{plotDistribution}
\title{plotDistribution}
\usage{
plotDistribution(object, sampling = "original", factor = NULL)
}
\arguments{
\item{object}{Cookie object}

\item{sampling}{choose original data or a specific sampling subset for analysis}

\item{factor}{The factor for plot}
}
\description{
Plot coverage on each factor for a sampling
}
