% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotReduction}
\alias{plotReduction}
\title{plotReduction}
\usage{
plotReduction(object, reduction = NULL, group.factor = NULL,
  sampling = NULL)
}
\arguments{
\item{object}{Cookie object}

\item{reduction}{Reduction results used.}

\item{group.factor}{Color samples by which factor.}

\item{sampling}{Hightlight the selected sample in a speficic sampling.}
}
\description{
Plot all samples on a 2D reduction, samples can be colored by factors.
}
