% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSizeTest}
\alias{plotSizeTest}
\title{plotSizeTest}
\usage{
plotSizeTest(object, type = "line", test.name = NULL)
}
\arguments{
\item{object}{Cookie object}

\item{type}{Figure type, could be "radar" or "line"}

\item{test.name}{The name for a size test}
}
\description{
Plot coverage on each factor for a size test to determine an appropriate sample size
}
