% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{reductionMDS}
\alias{reductionMDS}
\title{reductionMDS}
\usage{
reductionMDS(object, eig = FALSE, k = 2, add = FALSE,
  x.ret = FALSE)
}
\arguments{
\item{object}{Cookie object}

\item{eig}{Indicates whether eigenvalues should be returned.}

\item{k}{The maximum dimension of the space which the data are to be represented in; must be in {1, 2, …, n-1}.}

\item{add}{Logical indicating if an additive constant c* should be computed, and added to the non-diagonal dissimilarities such that the modified dissimilarities are Euclidean.}

\item{x.ret}{Indicates whether the doubly centred symmetric distance matrix should be returned.}
}
\description{
Run Classical MDS to project samples into 2D space using pairwise distances
}
