% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{reductionTSNE}
\alias{reductionTSNE}
\title{reductionTSNE}
\usage{
reductionTSNE(object, perplexity = 30, dim = 2, seed = 42,
  theta = 0.5)
}
\arguments{
\item{object}{(For Seurat) Seurat object}

\item{perplexity}{numeric; Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation)}

\item{dim}{integer; Output dimensionality (default: 2)}

\item{seed}{integer; seed for reproducible results.}

\item{theta}{numeric; Speed/accuracy trade-off (increase for less accuracy), set to 0.0 for exact TSNE (default: 0.5)}

\item{assay}{(For Seurat) run t-SNE for which assay, choose from RNA, ADT, Joint or All}
}
\description{
Run t-SNE to project samples into 2D map using pairwise distances
}
