% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{SampleSizeTest-class}
\alias{SampleSizeTest-class}
\alias{SampleSizeTest}
\title{The Sample Size Test Class}
\description{
The SampleSizeTest object stores a dimension reduction result generated by t-SNE, UMAP or MDS
}
\section{Slots}{

\describe{
\item{\code{prime.factor}}{Prime factor (if determined)}

\item{\code{coverage}}{Coverage on each factor under different sample size}

\item{\code{selection}}{Sample selected under different sample size}
}}

