% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{sampleSizeTest}
\alias{sampleSizeTest}
\title{sampleSizeTest}
\usage{
sampleSizeTest(object = NULL, prime.factor = NULL, size.range = NULL,
  name = NULL)
}
\arguments{
\item{object}{Cookie object}

\item{prime.factor}{The unique prime factor.}

\item{size.range}{Sample size range}

\item{name}{A name for this run. e.g. test1}
}
\description{
Run sample size test for current dataset
}
