% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{Sampling-class}
\alias{Sampling-class}
\alias{Sampling}
\title{The Sampling Class}
\description{
The Sampling object stores a dimension reduction result generated by t-SNE, UMAP or MDS
}
\section{Slots}{

\describe{
\item{\code{prime.factor}}{Prime factor (if determined)}

\item{\code{important.factor}}{Important factor}

\item{\code{size}}{Size}

\item{\code{sampling}}{Sample selected under different sample size}
}}

