% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{sampling}
\alias{sampling}
\title{sampling}
\usage{
sampling(object = NULL, prime.factor = NULL, important.factor = NULL,
  sample.size = NULL, name = NULL)
}
\arguments{
\item{object}{Cookie object}

\item{prime.factor}{The unique prime factor.}

\item{important.factor}{the important factors}

\item{sample.size}{Sample size}

\item{name}{A name for this run. e.g. test1}
}
\description{
Sampling from current dataset
}
