% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{sign}
\alias{sign}
\title{Sign a Cookie Value}
\usage{
sign(val, secret)
}
\arguments{
\item{val}{A string. The cookie value to sign.}

\item{secret}{The secret key used to generate the signature.}
}
\value{
A string of the form \code{"<val>.<signature>"}.
}
\description{
Sign the given \code{val} with \code{secret}.
}
\examples{
sign("hello", "tobiiscool")

}
