% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{simpleRandomSampling}
\alias{simpleRandomSampling}
\title{simpleRandomSampling}
\usage{
simpleRandomSampling(object = NULL, sample.size = NULL, name = NULL)
}
\arguments{
\item{object}{Cookie object}

\item{sample.size}{Sample size}

\item{name}{A name for this run. e.g. test1}
}
\description{
Simple Random Sampling from current dataset
}
