% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{stratifiedSampling}
\alias{stratifiedSampling}
\title{stratifiedSampling}
\usage{
stratifiedSampling(object = NULL, prime.factor = NULL,
  sample.size = NULL, name = NULL)
}
\arguments{
\item{object}{Cookie object}

\item{prime.factor}{The unique prime factor.}

\item{sample.size}{Sample size}

\item{name}{A name for this run. e.g. test1}
}
\description{
Stratified Sampling from current dataset
}
