% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSampling}
\alias{plotSampling}
\title{plotSampling}
\usage{
plotSampling(object, name = NULL, coverage.type = "pct")
}
\arguments{
\item{object}{Cookie object}

\item{name}{The name for a size test}

\item{coverage.type}{users can choose "pct" or "cc"}
}
\description{
Plot coverage on each factor for a sampling
}
