% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSizeTest}
\alias{plotSizeTest}
\title{plotSizeTest}
\usage{
plotSizeTest(object, chart.type = "line", test.name = NULL,
  coverage.type = "pct")
}
\arguments{
\item{object}{Cookie object}

\item{chart.type}{Figure type, could be "radar" or "line"}

\item{test.name}{The name for a size test}

\item{coverage.type}{users can choose "pct" or "cc"}
}
\description{
Plot coverage on each factor for a size test to determine an appropriate sample size
}
